backend/requirements.txt
flask==2.3.3
flask_sqlalchemy==3.0.5
flask_cors==4.0.0
psycopg2-binary==2.9.7
SPARQLWrapper==2.0.0
rdflib==6.3.2
python-dotenv==1.0.0
marshmallow==3.20.1
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
requests==2.31.0
